Add Sass and create pages directory
