+++ package.json
@@ -6,13 +6,14 @@
   "dependencies": {
     "axios": "^0.18.0",
     "d3": "^4.13.0",
+    "node-sass": "^4.9.3",
     "npm": "^6.0.0",
     "react": "^16.2.0",
     "react-cookies": "^0.1.0",
     "react-dom": "^16.2.0",
     "react-helmet": "^5.2.0",
     "react-redux": "^5.0.7",
-    "react-scripts": "1.1.0",
+    "react-scripts": "2.0.4",
     "react-scroll-to-component": "^1.0.2",
     "react-select": "^1.2.1",
     "redux": "^3.7.2",
@@ -29,5 +30,11 @@
   },
   "devDependencies": {
     "gh-pages": "^1.1.0"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
