+++ src/index.css
@@ -1,5 +1,20 @@
-body {
+@import url('https://rsms.me/inter/inter-ui.css');
+
+html, body {
   margin: 0;
   padding: 0;
-  font-family: sans-serif;
+  font-family: "Inter UI", sans-serif;
+  font-size: 16px;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  overflow-x: hidden;
+}
+
+.root {
+  display: flex;
+  flex-flow: column nowrap;
+  height: 100vh;
+  width: 100vw;
+  overflow-x: hidden;
+  overflow-y: auto;
 }
