+++ src/pages/App.scss
@@ -1,0 +1,2 @@
+.app {
+}
